chore(manifests): move utility methods in KustomizeTemplateUtils to BakeManifestEnvironment
to pave the way for helm to use git/repo artifacts
