+++ rosco-manifests/src/main/java/com/netflix/spinnaker/rosco/manifests/kustomize/KustomizeTemplateUtils.java
@@ -22,10 +22,8 @@ import com.netflix.spinnaker.rosco.jobs.BakeRecipe;
 import com.netflix.spinnaker.rosco.manifests.ArtifactDownloader;
 import com.netflix.spinnaker.rosco.manifests.BakeManifestEnvironment;
 import com.netflix.spinnaker.rosco.manifests.kustomize.mapping.Kustomization;
-import java.io.BufferedInputStream;
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
 import java.net.URI;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -37,9 +35,6 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.compress.archivers.ArchiveEntry;
-import org.apache.commons.compress.archivers.tar.TarArchiveInputStream;
-import org.apache.commons.compress.compressors.gzip.GzipCompressorInputStream;
 import org.apache.commons.io.FilenameUtils;
 import org.springframework.stereotype.Component;
 
@@ -120,18 +115,7 @@ public class KustomizeTemplateUtils {
       throw new IllegalArgumentException("The bake request should contain a kustomize file path.");
     }
 
-    InputStream inputStream;
-    try {
-      inputStream = artifactDownloader.downloadArtifact(artifact);
-    } catch (IOException e) {
-      throw new IOException("Failed to download git/repo artifact: " + e.getMessage(), e);
-    }
-
-    try {
-      extractArtifact(inputStream, env.resolvePath(""));
-    } catch (IOException e) {
-      throw new IOException("Failed to extract git/repo artifact: " + e.getMessage(), e);
-    }
+    env.downloadArtifactTarballAndExtract(artifactDownloader, artifact);
 
     List<String> command = new ArrayList<>();
     command.add("kustomize");
@@ -141,34 +125,6 @@ public class KustomizeTemplateUtils {
     BakeRecipe result = new BakeRecipe();
     result.setCommand(command);
     return result;
-  }
-
-  // This being here is temporary until we find a better way to abstract it
-  private static void extractArtifact(InputStream inputStream, Path outputPath) throws IOException {
-    try (TarArchiveInputStream tarArchiveInputStream =
-        new TarArchiveInputStream(
-            new GzipCompressorInputStream(new BufferedInputStream(inputStream)))) {
-
-      ArchiveEntry archiveEntry;
-      while ((archiveEntry = tarArchiveInputStream.getNextEntry()) != null) {
-        Path archiveEntryOutput = validateArchiveEntry(archiveEntry.getName(), outputPath);
-        if (archiveEntry.isDirectory()) {
-          if (!Files.exists(archiveEntryOutput)) {
-            Files.createDirectory(archiveEntryOutput);
-          }
-        } else {
-          Files.copy(tarArchiveInputStream, archiveEntryOutput);
-        }
-      }
-    }
-  }
-
-  private static Path validateArchiveEntry(String archiveEntryName, Path outputPath) {
-    Path entryPath = outputPath.resolve(archiveEntryName);
-    if (!entryPath.normalize().startsWith(outputPath)) {
-      throw new IllegalStateException("Attempting to create a file outside of the staging path.");
-    }
-    return entryPath;
   }
 
   protected void downloadArtifactToTmpFileStructure(
